Rename to selection_panel.rs
