+++ emergence_lib/src/ui/selection_panel.rs
